Fixed getAPM.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@3395 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/include/BWAPI/Client/GameData.h
@@ -27,8 +27,8 @@ namespace BWAPI
     bool isDebug;
 
     int instanceID;
-    double botAPM_noselects;
-    double botAPM_selects;
+    int botAPM_noselects;
+    int botAPM_selects;
 
     //forces
     int forceCount;
