+++ bwapi/include/BWAPI/Client/GameImpl.h
@@ -215,7 +215,7 @@ namespace BWAPI
       virtual int  getReplayFrameCount();
       virtual void setGUI(bool enabled = true);
       virtual int  getInstanceNumber();
-      virtual double getAPM(bool includeSelects = false);
+      virtual int getAPM(bool includeSelects = false);
       virtual bool setMap(const char *mapFileName);
   };
 }
